Model field fixed

+++ api/models.py
@@ -18,7 +18,7 @@ class Book(models.Model):
     authors = models.ManyToManyField(Author, related_name="authors", blank=True)
     published_date = models.CharField(max_length=32)
     categories = models.ManyToManyField(Category, related_name="category", blank=True)
-    average_ratings = models.IntegerField(null=True, blank=True)
+    average_ratings = models.FloatField(null=True, blank=True)
     ratings_count = models.IntegerField(null=True, blank=True)
     thumbnail = models.CharField(max_length=256, default='')
